start create account refactor
