+++ app/src/main/java/com/group80/uoftinder/presenter_layer/create_account/CreateAccountPresenter.java
@@ -1,5 +1,7 @@
-package com.group80.uoftinder.create_account_use_case;
+package com.group80.uoftinder.presenter_layer.create_account;
 
+import com.group80.uoftinder.create_account_use_case.CreateAccountPresenterInterface;
+import com.group80.uoftinder.create_account_use_case.CreateAccountViewInterface;
 import com.group80.uoftinder.entities_layer.User;
 
 // Interface adapter layer
